refactor(game): use Unit type instead of hardcoded f64

+++ src/utils/time.odin
@@ -3,6 +3,6 @@ package utils
 import "core:time"
 
 // Convert float seconds to time.Duration seconds
-f2s :: proc(t: f64) -> time.Duration {
-	return time.Duration(t * f64(time.Second))
+f2s :: proc(t: f32) -> time.Duration {
+	return time.Duration(t * f32(time.Second))
 }
